xen/grant: Fix signed/unsigned comparisons issues
Each function takes an unsigned count, and loops based on a signed i.  This
causes problems when between 2 and 4 billion operations are requested.

In practice, signed-ness issues aren't possible because count exceeding
INT_MAX is excluded earlier in do_grant_op(), but the code reads as if it is
buggy, and GCC obviously can't spot this either.

Bloat-o-meter reports:
  add/remove: 0/0 grow/shrink: 0/4 up/down: 0/-95 (-95)
  Function                                     old     new   delta
  do_grant_table_op                           7155    7140     -15
  gnttab_transfer                             2732    2716     -16
  gnttab_unmap_grant_ref                       771     739     -32
  gnttab_unmap_and_replace                     771     739     -32
  Total: Before=2996364, After=2996269, chg -0.00%

and inspection of gnttab_unmap_grant_ref() at least reveals one fewer local
variables on the stack.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <20200227184602.28282-1-andrew.cooper3@citrix.com>
Reviewed-by: Jan Beulich <jbeulich@suse.com>