scripts: Use stat to check lock claim
Replace the perl locking check with stat(1).  Stat is able to fstat
stdin (file descriptor 0) when passed '-' as an argument.  This is now
used to check $_lockfd.  stat(1) support for '-' was introduced to
coreutils in 2009.

After A releases its lock, script B will return from flock and execute
stat.  Since the lockfile has been removed from A, stat prints an error
to stderr and exits non-zero.  '|| :' is needed to squash the non-zero
exit status - otherwise the script terminates since `set -e` is enabled.
stderr needs to be redirected to /dev/null otherwise
/var/log/xen/xen-hotplug.log will get filled with "No such file or
directory" messages.

This change removes the only runtime dependency of the xen toolstack on
perl.

While here, replace some tabs with spaces to match the rest of the file.

Suggested-by: Ian Jackson <ian.jackson@citrix.com>
Signed-off-by: Jason Andryuk <jandryuk@gmail.com>
Message-Id: <20200310132904.16992-1-jandryuk@gmail.com>
Acked-by: Wei Liu <wl@xen.org>