x86/ucode: Alter ops->free_patch() to free the entire patch
The data layout for struct microcode_patch is extremely poor, and
unnecessarily complicated.  Almost all of it is opaque to core.c, with the
exception of free_patch().

Move the responsibility for freeing the patch into the free_patch() hook,
which will allow each driver to do a better job.  Take the opportunity to make
the hooks idempotent.

No practical change in behaviour.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>
Message-Id: <20200319152622.31758-6-andrew.cooper3@citrix.com>