x86/nvmx: clarify and fix usage of nvmx_update_apicv
The current usage of nvmx_update_apicv is not clear: it is deeply
intertwined with the Ack interrupt on exit VMEXIT control.

The code in nvmx_update_apicv should update the SVI (in service interrupt)
field of the guest interrupt status only when the Ack interrupt on
exit is set, as it is used to record that the interrupt being
serviced is signaled in a vmcs field, and hence hasn't been injected
as on native. It's important to record the current in service
interrupt on the guest interrupt status field, or else further
interrupts won't respect the priority of the in service one.

While clarifying the usage make sure that the SVI is only updated when
Ack on exit is set and the nested vmcs interrupt info field is valid. Or
else a guest not using the Ack on exit feature would loose interrupts as
they would be signaled as being in service on the guest interrupt
status field but won't actually be recorded on the interrupt info vmcs
field, neither injected in any other way.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Message-Id: <20200320190737.42110-3-roger.pau@citrix.com>