xen/mm: Move the MM types in a separate header
It is getting incredibly difficult to use typesafe GFN/MFN/PFN in the
headers because of circular dependency. For instance, asm-x86/page.h
cannot include xen/mm.h.

In order to convert more code to use typesafe, the types are now moved
in a separate header that requires only a few dependencies.

Signed-off-by: Julien Grall <jgrall@amazon.com>
Message-Id: <20200322161418.31606-4-julien@xen.org>