xen: Convert virt_to_mfn() and mfn_to_virt() to use typesafe MFN
Most of Xen is now either override the helpers virt_to_mfn() and
mfn_to_virt() to use typesafe MFN or use mfn_x() to remove the
typesafety when calling the helpers.

Therefore it is time to switch the two helpers to use typesafe MFN and
remove completely the possibly to make them unsafe by dropping the
double-underscore version.

Places that were still using non-typesafe MFN have been either
converted to use typesafe (if the changes are simple) or use
_mfn()/mfn_x() until the rest of the code is changed.

There are a couple of noticeable changes in the code:
    - pvh_populate_p2m() were storing the mfn in a variable called
      'addr'. This has now been renamed to 'mfn'.
    - allocate_cache_aligned_memnodemap() were storing an address in a
      variable called 'mfn'. The code has been reworked to avoid
      repurposing the variable.

No functional changes intended.

Signed-off-by: Julien Grall <jgrall@amazon.com>
Message-Id: <20200322161418.31606-5-julien@xen.org>