xen: x86: make init_intel_cacheinfo() void.
It seems that we took this code from Linux, back when the function was
'unsigned int' and the return value was used.

But we are currently not doing anything with such value, so let's get
rid of it and make the function void. As an anecdote, that's pretty much
the same that happened in Linux as, since commit 807e9bc8e2fe6 ("x86/CPU:
Move cpu_detect_cache_sizes() into init_intel_cacheinfo()") the function
is void there too.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>
Message-Id: <158524305235.31663.14043511848631384508.stgit@Palanthas>
Reviewed-by: Andrew Cooper <andrew.cooper3@citrix.com>